refactor: move client routes to container folder
