+++ apps/client/src/app/$group-id/$transaction-id/edit.tsx
@@ -1,10 +1,1 @@
-import { Header } from '@/components/haeder'
-import { createFileRoute } from '@tanstack/react-router'
-
-export const Route = createFileRoute('/$group-id/$transaction-id/edit')({
-  component: RouteComponent
-})
-
-function RouteComponent(): JSX.Element {
-  return <Header title="Editar transação" hasBackButton />
-}
+export { TransactionIdEditRoute as Route } from '@/containers/groups/transaction-id-edit.route'
