+++ apps/client/src/app/activities/index.tsx
@@ -1,15 +1,1 @@
-import { Header } from '@/components/haeder'
-import { createFileRoute } from '@tanstack/react-router'
-
-export const Route = createFileRoute('/activities/')({
-  component: RouteComponent
-})
-
-function RouteComponent(): JSX.Element {
-  return (
-    <>
-      <Header title="Atividades" />
-      'Hello /activities/!'
-    </>
-  )
-}
+export { ActivitiesRoute as Route } from '@/containers/activities/activities.route'
