+++ apps/client/src/app/friends/index.tsx
@@ -1,15 +1,1 @@
-import { Header } from '@/components/haeder'
-import { createFileRoute } from '@tanstack/react-router'
-
-export const Route = createFileRoute('/friends/')({
-  component: RouteComponent
-})
-
-function RouteComponent(): JSX.Element {
-  return (
-    <>
-      <Header title="Amigos" />
-      'Hello /friends/!'
-    </>
-  )
-}
+export { FriendsRoute as Route } from '@/containers/friends/friends.route'
