+++ apps/client/src/app/new-transaction/index.tsx
@@ -1,0 +1,1 @@
+export { NewTransactionRoute as Route } from '@/containers/new-transaction/new-transaction.route'
