+++ apps/client/src/route-tree.gen.ts
@@ -12,8 +12,8 @@
 
 import { Route as rootRoute } from './app/__root'
 import { Route as IndexImport } from './app/index'
-import { Route as TransactionIndexImport } from './app/transaction/index'
 import { Route as SignInIndexImport } from './app/sign-in/index'
+import { Route as NewTransactionIndexImport } from './app/new-transaction/index'
 import { Route as FriendsIndexImport } from './app/friends/index'
 import { Route as ActivitiesIndexImport } from './app/activities/index'
 import { Route as AccountIndexImport } from './app/account/index'
@@ -29,15 +29,15 @@ const IndexRoute = IndexImport.update({
   getParentRoute: () => rootRoute,
 } as any)
 
-const TransactionIndexRoute = TransactionIndexImport.update({
-  id: '/transaction/',
-  path: '/transaction/',
-  getParentRoute: () => rootRoute,
-} as any)
-
 const SignInIndexRoute = SignInIndexImport.update({
   id: '/sign-in/',
   path: '/sign-in/',
+  getParentRoute: () => rootRoute,
+} as any)
+
+const NewTransactionIndexRoute = NewTransactionIndexImport.update({
+  id: '/new-transaction/',
+  path: '/new-transaction/',
   getParentRoute: () => rootRoute,
 } as any)
 
@@ -116,18 +116,18 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof FriendsIndexImport
       parentRoute: typeof rootRoute
     }
+    '/new-transaction/': {
+      id: '/new-transaction/'
+      path: '/new-transaction'
+      fullPath: '/new-transaction'
+      preLoaderRoute: typeof NewTransactionIndexImport
+      parentRoute: typeof rootRoute
+    }
     '/sign-in/': {
       id: '/sign-in/'
       path: '/sign-in'
       fullPath: '/sign-in'
       preLoaderRoute: typeof SignInIndexImport
-      parentRoute: typeof rootRoute
-    }
-    '/transaction/': {
-      id: '/transaction/'
-      path: '/transaction'
-      fullPath: '/transaction'
-      preLoaderRoute: typeof TransactionIndexImport
       parentRoute: typeof rootRoute
     }
     '/$group-id/$transaction-id/edit': {
@@ -155,8 +155,8 @@ export interface FileRoutesByFullPath {
   '/account': typeof AccountIndexRoute
   '/activities': typeof ActivitiesIndexRoute
   '/friends': typeof FriendsIndexRoute
+  '/new-transaction': typeof NewTransactionIndexRoute
   '/sign-in': typeof SignInIndexRoute
-  '/transaction': typeof TransactionIndexRoute
   '/$group-id/$transaction-id/edit': typeof GroupIdTransactionIdEditRoute
   '/$group-id/$transaction-id': typeof GroupIdTransactionIdIndexRoute
 }
@@ -167,8 +167,8 @@ export interface FileRoutesByTo {
   '/account': typeof AccountIndexRoute
   '/activities': typeof ActivitiesIndexRoute
   '/friends': typeof FriendsIndexRoute
+  '/new-transaction': typeof NewTransactionIndexRoute
   '/sign-in': typeof SignInIndexRoute
-  '/transaction': typeof TransactionIndexRoute
   '/$group-id/$transaction-id/edit': typeof GroupIdTransactionIdEditRoute
   '/$group-id/$transaction-id': typeof GroupIdTransactionIdIndexRoute
 }
@@ -180,8 +180,8 @@ export interface FileRoutesById {
   '/account/': typeof AccountIndexRoute
   '/activities/': typeof ActivitiesIndexRoute
   '/friends/': typeof FriendsIndexRoute
+  '/new-transaction/': typeof NewTransactionIndexRoute
   '/sign-in/': typeof SignInIndexRoute
-  '/transaction/': typeof TransactionIndexRoute
   '/$group-id/$transaction-id/edit': typeof GroupIdTransactionIdEditRoute
   '/$group-id/$transaction-id/': typeof GroupIdTransactionIdIndexRoute
 }
@@ -194,8 +194,8 @@ export interface FileRouteTypes {
     | '/account'
     | '/activities'
     | '/friends'
+    | '/new-transaction'
     | '/sign-in'
-    | '/transaction'
     | '/$group-id/$transaction-id/edit'
     | '/$group-id/$transaction-id'
   fileRoutesByTo: FileRoutesByTo
@@ -205,8 +205,8 @@ export interface FileRouteTypes {
     | '/account'
     | '/activities'
     | '/friends'
+    | '/new-transaction'
     | '/sign-in'
-    | '/transaction'
     | '/$group-id/$transaction-id/edit'
     | '/$group-id/$transaction-id'
   id:
@@ -216,8 +216,8 @@ export interface FileRouteTypes {
     | '/account/'
     | '/activities/'
     | '/friends/'
+    | '/new-transaction/'
     | '/sign-in/'
-    | '/transaction/'
     | '/$group-id/$transaction-id/edit'
     | '/$group-id/$transaction-id/'
   fileRoutesById: FileRoutesById
@@ -229,8 +229,8 @@ export interface RootRouteChildren {
   AccountIndexRoute: typeof AccountIndexRoute
   ActivitiesIndexRoute: typeof ActivitiesIndexRoute
   FriendsIndexRoute: typeof FriendsIndexRoute
+  NewTransactionIndexRoute: typeof NewTransactionIndexRoute
   SignInIndexRoute: typeof SignInIndexRoute
-  TransactionIndexRoute: typeof TransactionIndexRoute
   GroupIdTransactionIdEditRoute: typeof GroupIdTransactionIdEditRoute
   GroupIdTransactionIdIndexRoute: typeof GroupIdTransactionIdIndexRoute
 }
@@ -241,8 +241,8 @@ const rootRouteChildren: RootRouteChildren = {
   AccountIndexRoute: AccountIndexRoute,
   ActivitiesIndexRoute: ActivitiesIndexRoute,
   FriendsIndexRoute: FriendsIndexRoute,
+  NewTransactionIndexRoute: NewTransactionIndexRoute,
   SignInIndexRoute: SignInIndexRoute,
-  TransactionIndexRoute: TransactionIndexRoute,
   GroupIdTransactionIdEditRoute: GroupIdTransactionIdEditRoute,
   GroupIdTransactionIdIndexRoute: GroupIdTransactionIdIndexRoute,
 }
@@ -262,8 +262,8 @@ export const routeTree = rootRoute
         "/account/",
         "/activities/",
         "/friends/",
+        "/new-transaction/",
         "/sign-in/",
-        "/transaction/",
         "/$group-id/$transaction-id/edit",
         "/$group-id/$transaction-id/"
       ]
@@ -283,11 +283,11 @@ export const routeTree = rootRoute
     "/friends/": {
       "filePath": "friends/index.tsx"
     },
+    "/new-transaction/": {
+      "filePath": "new-transaction/index.tsx"
+    },
     "/sign-in/": {
       "filePath": "sign-in/index.tsx"
-    },
-    "/transaction/": {
-      "filePath": "transaction/index.tsx"
     },
     "/$group-id/$transaction-id/edit": {
       "filePath": "$group-id/$transaction-id/edit.tsx"
